Refactor to make comp_organizer_heatlist work
